Move virtual api and storage into pxtlib
