+++ webapp/src/core.ts
@@ -299,7 +299,7 @@ export const ENTER_KEY = 13;
 export const SPACE_KEY = 32;
 
 export function getHighContrastOnce(): boolean {
-    return data.getData<boolean>(auth.HIGHCONTRAST) || false
+    return pxt.data.getData<boolean>(auth.HIGHCONTRAST) || false
 }
 export function toggleHighContrast() {
     setHighContrast(!getHighContrastOnce())
@@ -337,28 +337,6 @@ export function findChild(c: React.Component<any, any>, selector: string): Eleme
     let self = ReactDOM.findDOMNode(c);
     if (!selector) return [self]
     return pxt.Util.toArray(self.querySelectorAll(selector));
-}
-
-export function parseQueryString(qs: string) {
-    let r: pxt.Map<string> = {}
-
-    qs.replace(/\+/g, " ").replace(/([^#?&=]+)=([^#?&=]*)/g, (f: string, k: string, v: string) => {
-        r[decodeURIComponent(k)] = decodeURIComponent(v)
-        return ""
-    })
-    return r
-}
-
-export function stringifyQueryString(url: string, qs: any) {
-    for (let k of Object.keys(qs)) {
-        if (url.indexOf("?") >= 0) {
-            url += "&"
-        } else {
-            url += "?"
-        }
-        url += encodeURIComponent(k) + "=" + encodeURIComponent(qs[k])
-    }
-    return url
 }
 
 export function handleNetworkError(e: any, ignoredCodes?: number[]) {
