+++ webapp/src/workspace.ts
@@ -223,8 +223,8 @@ function refreshHeadersSession() {
     if (isHeadersSessionOutdated()) {
         pxt.storage.setLocal('workspacesessionid', sessionID);
         pxt.debug(`workspace: refreshed headers session to ${sessionID}`);
-        data.invalidate("header:*");
-        data.invalidate("text:*");
+        pxt.data.invalidate("header:*");
+        pxt.data.invalidate("text:*");
     }
 }
 // this is an identifier for the current frame
@@ -574,13 +574,13 @@ export async function saveAsync(h: Header, text?: ScriptText, fromCloudSync?: bo
         }
 
         if (text || h.isDeleted) {
-            data.invalidate("text:" + h.id);
-            data.invalidate("pkg-git-status:" + h.id);
+            pxt.data.invalidate("text:" + h.id);
+            pxt.data.invalidate("pkg-git-status:" + h.id);
         }
         data.invalidateHeader("header", h);
         if (newSave) {
             // the count of headers has changed
-            data.invalidate("headers:");
+            pxt.data.invalidate("headers:");
         }
 
         refreshHeadersSession();
@@ -950,7 +950,7 @@ export async function commitAsync(hd: Header, options: CommitOptions = {}) {
     if (newCommit == null) {
         return commitId
     } else {
-        data.invalidate("gh-commits:*"); // invalid any cached commits
+        pxt.data.invalidate("gh-commits:*"); // invalid any cached commits
         // if we created a block preview, add as comment
         if (blocksDiffSha) {
             await pxt.github.postCommitComment(
@@ -1548,7 +1548,7 @@ export function syncAsync(): Promise<pxt.editor.EditorSyncState> {
                         data.invalidateHeader("header", hd);
                         data.invalidateHeader("text", hd);
                         data.invalidateHeader("pkg-git-status", hd);
-                        data.invalidate("gh-commits:*"); // invalidate commits just in case
+                        pxt.data.invalidate("gh-commits:*"); // invalidate commits just in case
                     }
                 } else {
                     ex = {
@@ -1580,7 +1580,7 @@ export function resetAsync() {
         .then(compiler.clearApiInfoDbAsync)
         .then(() => {
             pxt.storage.clearLocal();
-            data.clearCache();
+            pxt.data.clearCache();
             // keep local token (localhost and electron) on reset
             if (Cloud.localToken)
                 pxt.storage.setLocal("local_token", Cloud.localToken);
@@ -1640,9 +1640,9 @@ export function dbgHdrToString(h: Header): string {
     header:*        - all headers
 */
 
-data.mountVirtualApi("header", {
+pxt.data.mountVirtualApi("header", {
     getSync: p => {
-        p = data.stripProtocol(p)
+        p = pxt.data.stripProtocol(p)
         if (p == "*") return getHeaders()
         return getHeader(p)
     },
@@ -1652,9 +1652,9 @@ data.mountVirtualApi("header", {
     headers:SEARCH   - search headers
 */
 
-data.mountVirtualApi("headers", {
+pxt.data.mountVirtualApi("headers", {
     getAsync: p => {
-        p = data.stripProtocol(p)
+        p = pxt.data.stripProtocol(p)
         const headers = getHeaders()
         if (!p) return Promise.resolve(headers)
         return compiler.projectSearchAsync({ term: p, headers })
@@ -1674,7 +1674,7 @@ data.mountVirtualApi("headers", {
     text:<guid>            - all files
     text:<guid>/<filename> - one file
 */
-data.mountVirtualApi("text", {
+pxt.data.mountVirtualApi("text", {
     getAsync: p => {
         const m = /^[\w\-]+:([^\/]+)(\/(.*))?/.exec(p)
         return getTextAsync(m[1])
